requirements.txt
gTTS
playsound
SpeechRecognition pydub
pyaudio
plyer
randfacts
win32api
screeninfo
Pillow